# cmd/apiserver/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AI Receipt Tracker read-only receipts API",
    "version": "1.0.0",
    "description": "Read-only receipt access for AI clients. Authenticate every operation with the server-issued bearer token in the form `Authorization: Bearer rk_ai_<key-id>.<secret>`. The token is scoped to its owner's receipts. The image operation returns JPEG bytes; it does not expose a Google Cloud Storage URL."
  },
  "servers": [
    {
      "url": "https://ai-receipt-tracker-backend-267658267276.northamerica-northeast2.run.app"
    }
  ],
  "tags": [
    {
      "name": "AI receipts",
      "description": "Read-only receipt data and image access for the authenticated AI token owner."
    }
  ],
  "paths": {
    "/ai/receipts": {
      "get": {
        "tags": ["AI receipts"],
        "operationId": "listAIReceipts",
        "summary": "List the owner's receipt summaries",
        "description": "Returns all receipt summaries owned by the AI bearer token's user, newest first.",
        "security": [
          {
            "AIAccessToken": []
          }
        ],
        "responses": {
          "200": {
            "description": "Receipt summaries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIReceiptSummaryList"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/ai/receipts/{receipt_id}": {
      "get": {
        "tags": ["AI receipts"],
        "operationId": "getAIReceipt",
        "summary": "Get one complete receipt",
        "description": "Returns the complete structured receipt record when the receipt belongs to the AI bearer token's user.",
        "security": [
          {
            "AIAccessToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReceiptID"
          }
        ],
        "responses": {
          "200": {
            "description": "Receipt record",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AIReceipt"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    },
    "/ai/receipts/{receipt_id}/image": {
      "get": {
        "tags": ["AI receipts"],
        "operationId": "getAIReceiptImage",
        "summary": "Get one receipt image as JPEG",
        "description": "Returns the owner's private stored receipt image converted to JPEG. The response contains image bytes, not a permanent or signed storage URL.",
        "security": [
          {
            "AIAccessToken": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/ReceiptID"
          }
        ],
        "responses": {
          "200": {
            "description": "Receipt image",
            "headers": {
              "Cache-Control": {
                "description": "The backend returns no-store for image responses.",
                "schema": {
                  "type": "string",
                  "example": "no-store"
                }
              },
              "X-Content-Type-Options": {
                "description": "The backend returns nosniff for image responses.",
                "schema": {
                  "type": "string",
                  "example": "nosniff"
                }
              }
            },
            "content": {
              "image/jpeg": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "AIAccessToken": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "rk_ai_<key-id>.<secret>",
        "description": "Use the server-issued AI access token exactly as `Authorization: Bearer rk_ai_<key-id>.<secret>`. Do not use the Google OAuth token or include the token in a URL."
      }
    },
    "parameters": {
      "ReceiptID": {
        "name": "receipt_id",
        "in": "path",
        "required": true,
        "description": "The Receipt Keeper receipt ID.",
        "schema": {
          "type": "string",
          "minLength": 1
        }
      }
    },
    "responses": {
      "Unauthorized": {
        "description": "Missing, malformed, revoked, or invalid AI access token.",
        "headers": {
          "WWW-Authenticate": {
            "description": "Bearer authentication is required.",
            "schema": {
              "type": "string",
              "example": "Bearer"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "The receipt is not owned by the token owner, or the receipt/image does not exist.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "The endpoint accepts GET only.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "The stored image dimensions are too large to convert safely.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "ServerError": {
        "description": "Unexpected backend or storage failure.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["detail"],
        "properties": {
          "detail": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ReceiptItem": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string"
          },
          "quantity": {
            "type": "number",
            "format": "double"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
      "AIReceiptSummary": {
        "type": "object",
        "required": ["id"],
        "properties": {
          "id": {
            "type": "string"
          },
          "vendor": {
            "type": "string"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "type": "string"
          },
          "purchase_date": {
            "type": "string"
          },
          "invoice_id": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "AIReceiptSummaryList": {
        "type": "object",
        "required": ["receipts", "count"],
        "properties": {
          "receipts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AIReceiptSummary"
            }
          },
          "count": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        },
        "additionalProperties": false
      },
      "AIReceipt": {
        "type": "object",
        "required": ["id", "items", "image_url", "extracted_text", "extracted_fields", "created_at"],
        "properties": {
          "id": {
            "type": "string"
          },
          "vendor": {
            "type": "string"
          },
          "subtotal": {
            "type": "number",
            "format": "double"
          },
          "tax": {
            "type": "number",
            "format": "double"
          },
          "total": {
            "type": "number",
            "format": "double"
          },
          "category": {
            "type": "string"
          },
          "purchase_date": {
            "type": "string"
          },
          "invoice_id": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReceiptItem"
            }
          },
          "image_url": {
            "type": "string"
          },
          "extracted_text": {
            "type": "string"
          },
          "extracted_fields": {
            "type": "object",
            "additionalProperties": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      }
    }
  }
}